Rename pydis_core/utils/errors.py to pydis_core/utils/error_handling.py
