+++ pydis_core/utils/__init__.py
@@ -1,8 +1,8 @@
 """Useful utilities and tools for Discord bot development."""
 
 from pydis_core.utils import (
-    _monkey_patches, caching, channel, commands, cooldown, function, interactions, logging, members, paste_service,
-    regex, scheduling
+    _monkey_patches, caching, channel, commands, cooldown, error_handling, function, interactions, logging, members,
+    paste_service, regex, scheduling
 )
 from pydis_core.utils._extensions import unqualify
 
@@ -29,6 +29,7 @@ __all__ = [
     channel,
     commands,
     cooldown,
+    error_handling,
     function,
     interactions,
     logging,
